+++ src/app/app.module.ts
@@ -6,18 +6,21 @@ import { ProductsModule } from './components/products/products.module';
 import { CartModule } from './components/cart/cart.module';
 
 import { AppComponent } from './app.component';
+import { FirstComponent } from './components/first/first.component';
+import { SharedModule } from './shared/shared.module';
 
 @NgModule({
   declarations: [
-    AppComponent
+    AppComponent,
+    FirstComponent
   ],
   imports: [
     BrowserModule,
     AppRoutingModule,
     ProductsModule,
-    CartModule
+    CartModule,
+    SharedModule
   ],
-  providers: [],
   bootstrap: [AppComponent]
 })
 export class AppModule { }