+++ src/app/components/products/product-list/product-list.component.ts
@@ -14,6 +14,6 @@ export class ProductListComponent {
   ) { }
 
   onBuy(id: number): void {
-    this.cartService.addProductToCart(id);
+    this.cartService.addProduct(id);
   }
 }