+++ src/app/components/products/product/product.component.ts
@@ -1,6 +1,6 @@
 import { ChangeDetectionStrategy, Component, EventEmitter, Input, Output } from '@angular/core';
 
-import { ProductModel } from '../../../models/product.model';
+import { ProductModel } from '../../../models';
 
 @Component({
   selector: 'app-product',