+++ src/app/models/index.ts
@@ -1,0 +1,2 @@
+export * from './config-options.model';
+export * from './product.model';