+++ src/app/shared/services/index.ts
@@ -1,2 +1,7 @@
 export * from './cart.service';
 export * from './products.service';
+export * from './local-storage.service';
+export * from './config-options.service';
+export * from './generator.service';
+export * from './generator.factory';
+export * from './constants.service';