+++ src/app/shared/services/products.service.ts
@@ -1,6 +1,6 @@
 import { Injectable } from '@angular/core';
 import { productsMock } from '../../mocks/products.mock';
-import { ProductModel } from '../../models/product.model';
+import { ProductModel } from '../../models';
 
 @Injectable({
   providedIn: 'root'